Eliminate entire repo
The main icepack repo will be using pooch for all data fetching and
rasterio for all gridded data reading, eliminating the need for almost
all of the code in this repo. The only remaining data will be the
hand-drawn outlines of various glaciers that are used to generate
meshes, as well as some Python code for fixing up the outlines.